Feature(goalRoutes, goalController): Protect Goal Routes
-Limit user access of CRUD routes to a single user's goals
-Identify users by a JWT token, and filter goals by the user for update
and delete functionality

+++ backend/controllers/goalController.js
@@ -1,30 +1,42 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 // TODO: docs
 
 const getGoals = asyncHandler(async (req, res) => {
-	const goals = await Goal.find()
+	const goals = await Goal.find({ user: req.user.id })
 	res.status(200).json(goals)
 })
 
 const setGoal = asyncHandler(async (req, res) => {
-	if(!req.body.text) {
+	if (!req.body.text) {
 		res.status(400)
 		throw new Error('Please add a text field')
 	}
 	const goal = await Goal.create({
-		text: req.body.text
+		text: req.body.text,
+		user: req.user.id
 	})
 	res.status(200).json(goal)
 })
 
 const updateGoal = asyncHandler(async (req, res) => {
 	const goal = await Goal.findById(req.params.id)
-
-	if(!goal) {
+	if (!goal) {
 		res.status(400)
 		throw new Error('Goal not found')
+	}
+
+	const user = await User.findById(req.user.id)
+	if (!user) {
+		res.status(401)
+		throw new Error('User not found')
+	}
+	// Ensure that a user can only update their own goals
+	if (goal.user.toString() !== user.id) {
+		res.status(401)
+		throw new Error('User not authorized')
 	}
 
 	const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { new: true })
@@ -33,16 +45,25 @@ const updateGoal = asyncHandler(async (req, res) => {
 
 const deleteGoal = asyncHandler(async (req, res) => {
 	const goal = await Goal.findById(req.params.id)
-
-	if(!goal) {
+	if (!goal) {
 		res.status(400)
 		throw new Error('Goal not found')
+	}
+
+	const user = await User.findById(req.user.id)
+	if (!user) {
+		res.status(401)
+		throw new Error('User not found')
+	}
+	// Ensure that a user can only delete their own goals
+	if (goal.user.toString() !== user.id) {
+		res.status(401)
+		throw new Error('User not authorized')
 	}
 
 	await Goal.deleteOne({
 		_id: req.params.id
 	})
-
 	res.status(200).json({ id: req.params.id })
 })
